change args dir structure
