+++ packages/widgetbook/lib/next.dart
@@ -3,7 +3,9 @@ library widgetbook.next;
 
 import 'package:meta/meta.dart';
 
-export 'src/next/args.dart';
+export 'src/next/args/arg.dart';
+export 'src/next/args/story_args.dart';
+export 'src/next/args/string_arg.dart';
 export 'src/next/component.dart';
 export 'src/next/experimental_badge.dart';
 export 'src/next/meta.dart';
